requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
openai==1.29.0
tenacity==8.2.2
aiofiles==0.8.0